[rename] delivery service interface package 정리
