+++ delivery/src/main/java/com/example/delivery/application/service/DeliveryQueryServiceImpl.java
@@ -1,0 +1,11 @@
+package com.example.delivery.application.service;
+
+import org.springframework.stereotype.Service;
+
+import lombok.RequiredArgsConstructor;
+
+@Service
+@RequiredArgsConstructor
+public class DeliveryQueryServiceImpl implements DeliveryQueryService{
+
+}
