+++ delivery/src/main/java/com/example/delivery/application/service/interfaces/DeliveryCreateService.java
@@ -1,4 +1,4 @@
-package com.example.delivery.application.service;
+package com.example.delivery.application.service.interfaces;
 
 import com.example.delivery.application.dto.CreateDeliveryResponseDto;
 import com.example.delivery.presentation.dto.CreateDeliveryRequestDto;
