+++ delivery/src/main/java/com/example/delivery/application/service/interfaces/DeliveryService.java
@@ -1,4 +1,4 @@
-package com.example.delivery.application.service;
+package com.example.delivery.application.service.interfaces;
 
 public interface DeliveryService extends
 	DeliveryCreateService, DeliveryUpdateService, DeliveryStatusUpdateService, DeliveryCancelService {
